Implement items import and export
- Change spoonWeight and sugarPercentage to doubles for better Json support
- Add formating for weight and sugarPercentage of ingredients
- Implement intent activity for file selection when importing and exporting
- Implement file reading and writing of itemsJsonString

+++ app/src/main/java/com/romanbrunner/apps/mueslirandomizer/Article.java
@@ -8,16 +8,16 @@ public interface Article
     String getName();
     String getBrand();
     Type getType();
-    float getSpoonWeight();
-    float getSugarPercentage();
+    double getSpoonWeight();
+    double getSugarPercentage();
     int getMultiplier();
     int getSelectionsLeft();
     boolean getIsFiller();
 
     void setName(String name);
     void setBrand(String brand);
-    void setSpoonWeight(float spoonWeight);
-    void setSugarPercentage(float sugarPercentage);
+    void setSpoonWeight(double spoonWeight);
+    void setSugarPercentage(double sugarPercentage);
     void setMultiplier(int multiplier);
     void setSelectionsLeft(int selectionsLeft);
     void setIsFiller(boolean isFiller);
